+++ frontend/src/components/SubmissionResult/Loading.tsx
