+++ frontend/src/components/SubmissionResult/ResultTotalInfo.tsx
