+++ frontend/src/components/SubmissionResult/index.tsx
