+++ frontend/src/components/SubmissionResult/types.ts
